+++ src/App.vue
@@ -1,72 +1,45 @@
 <script setup lang="ts">
-import { inject, ref, onMounted } from "vue";
+import { inject, ref, unref, onMounted, watch } from "vue";
+import { I18n } from "vue-i18n";
 
 import Breadcrumb from "./components/Breadcrumb.vue";
+import Export from "./components/Export.vue";
 import Form from "./components/Form.vue";
 
 import { Client } from "./client/Client";
 import { notify } from "./utils/notify";
-import { dump } from "./utils/export";
+import { dumpForm } from "./utils/export";
+import { getArcoLang } from "./utils/language";
 
 import { ISiyuan } from "./types/siyuan/siyuan";
-import { IData } from "./types/data";
+import { IData, IAL } from "./types/data";
 import { IForm } from "./types/form";
 
 const client = inject("client") as Client;
 const siyuan = inject("siyuan") as ISiyuan;
 const data = inject("data") as IData;
-const editable = ref(false);
+const i18n = inject("i18n") as I18n;
+
+const editable = ref(false); // 是否可编辑
+const active_key_collapse = ref([1, 2]); // 属性编辑页签展开的列表
+const active_key_tags = ref(1); // 选中的页签
+const arch_lang = getArcoLang(unref(i18n.global.locale));
+
+watch(active_key_tags, updated, { flush: "post" });
 
 /* 更新文档 */
-function updated(form?: IForm): void {
+function updated(): void {
     setTimeout(() => {
+        if (import.meta.env.DEV) {
+            console.log("APP.updated");
+        }
+
         /* 调整高度 */
         const height = document.getElementById("main")?.scrollHeight;
         if (height && data.element) {
             data.element.style.height = `${height + 2}px`;
         }
-
-        /* 保存元数据 */
-        // if (form) {
-        //     saveMetadata(form);
-        // }
     }, 0);
-}
-
-/* 保存元数据 */
-function saveMetadata(form: IForm): void {
-    const yaml = dump(form);
-
-    const metadata: string[] = [];
-
-    metadata.push("---");
-    metadata.push("\n");
-    metadata.push(yaml);
-    metadata.push("---");
-
-    const markdown = metadata.join("");
-    if (import.meta.env.DEV) {
-        console.log(markdown);
-    }
-
-    if (markdown !== data.block_ial["data-export-md"]) {
-        client
-            .setBlockAttrs({
-                id: data.block_id,
-                attrs: {
-                    "data-export-md": markdown,
-                },
-            })
-            .then(() => {
-                if (import.meta.env.DEV) {
-                    console.log(`Form.saveMetadata\n${markdown}`);
-                }
-            })
-            .catch(error => {
-                notify(error.toString());
-            });
-    }
-
 }
 
 /* 重新加载 */
@@ -84,61 +57,99 @@ onMounted(() => {
 </script>
 
 <template>
-    <a-layout
-        id="main"
-        class="layout"
-    >
-        <a-layout-header class="header">
-            <a-row>
-                <a-col flex="none">
-                    <Breadcrumb
-                        :paths="data.paths"
-                        :hpaths="data.hpaths"
-                    />
-                </a-col>
-                <a-col flex="auto">
-                    <div>auto</div>
-                </a-col>
-                <a-col flex="none">
-                    <a-button
-                        @click="reload"
-                        :title="$t('reload')"
-                        size="mini"
-                        type="outline"
-                    >
-                        <template #icon>
-                            <icon-refresh />
+    <a-config-provider :locale="arch_lang">
+        <a-layout
+            id="main"
+            class="layout"
+        >
+            <!-- 页眉 -->
+            <a-layout-header class="header">
+                <a-row>
+                    <a-col flex="none">
+                        <Breadcrumb
+                            :paths="data.paths"
+                            :hpaths="data.hpaths"
+                        />
+                    </a-col>
+                    <a-col flex="auto" />
+                    <a-col flex="none">
+                        <a-button
+                            @click="reload"
+                            :title="$t('reload')"
+                            size="mini"
+                            type="outline"
+                        >
+                            <template #icon>
+                                <icon-refresh />
+                            </template>
+                        </a-button>
+                        <a-divider
+                            class="divider-vertical"
+                            direction="vertical"
+                            margin="0.5em"
+                            :size="2"
+                        />
+                        <a-switch
+                            v-model:model-value="editable"
+                            :title="$t('editable')"
+                            style="margin-top: -4px"
+                            type="circle"
+                            size="medium"
+                        >
+                            <template #checked-icon>
+                                <icon-check />
+                            </template>
+                            <template #unchecked-icon>
+                                <icon-close />
+                            </template>
+                        </a-switch>
+                    </a-col>
+                </a-row>
+            </a-layout-header>
+
+            <!-- 主体 -->
+            <a-layout-content class="content">
+                <a-tabs
+                    v-model:active-key="active_key_tags"
+                    class="tags"
+                    size="mini"
+                    type="card-gutter"
+                >
+                    <!-- 属性编辑 -->
+                    <a-tab-pane :key="1">
+                        <template #title>
+                            <icon-edit />
+                            {{ $t("edit") }}
                         </template>
-                    </a-button>
-                    <a-divider direction="vertical" />
-                    <a-switch
-                        v-model:model-value="editable"
-                        :title="$t('editable')"
-                        style="margin-top: -4px"
-                        type="circle"
-                    >
-                        <template #checked-icon>
-                            <icon-check />
+
+                        <Form
+                            v-model:active-key="active_key_collapse"
+                            :client="client"
+                            :data="data"
+                            :editable="editable"
+                            @updated="updated"
+                        />
+                    </a-tab-pane>
+
+                    <!-- 属性导出 -->
+                    <a-tab-pane :key="2">
+                        <template #title>
+                            <icon-edit />
+                            {{ $t("export") }}
                         </template>
-                        <template #unchecked-icon>
-                            <icon-close />
-                        </template>
-                    </a-switch>
-                </a-col>
-            </a-row>
-        </a-layout-header>
-        <a-layout>
-            <a-layout-content>
-                <Form
-                    @updated="updated"
-                    :client="client"
-                    :data="data"
-                    :editable="editable"
-                />
+
+                        <Export
+                            :client="client"
+                            :data="data"
+                            :activate="active_key_tags === 2"
+                            @updated="updated"
+                        />
+                    </a-tab-pane>
+                </a-tabs>
             </a-layout-content>
+            <!-- <a-layout-footer>Footer</a-layout-footer> -->
         </a-layout>
-        <!-- <a-layout-footer>Footer</a-layout-footer> -->
-    </a-layout>
+    </a-config-provider>
 </template>
 
 <style scoped lang="less">
@@ -161,6 +172,20 @@ onMounted(() => {
     }
 
     :deep(.arco-layout-content) {
+    }
+
+    .divider-vertical {
+    }
+
+    /* 主体 */
+    .content {
+        /* 标签页 */
+        .tags {
+            /* 标签页内容 */
+            :deep(.arco-tabs-content) {
+                padding: 0;
+            }
+        }
     }
 }
 </style>